修正：障害物に当たったら時間を減らすスクリプトが機能していなかったのを修正した

+++ Assets/Scripts/GameManager.cs
@@ -89,7 +89,7 @@ public class GameManager : MonoBehaviour
         m_state = State.Finish;
     }
 
-    public void DecreaceTimer(int time)
+    public void DecreaceTimer(float time)
     {
         Timer -= time;
     }
